+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 
 - Implement serial workflows via "psimulate workflow" command
 - Add facade to interact with Jobmon SDK
+- Reduce simulation task runtime resource during restart with backups
 
 **3.1.4 - 05/18/26**